Add flushing to fileexporter
Add flush_interval in seconds to the fileexporter configuration
and use this parameter to periodically flush the buffer to the file.

Flushing eases debugging because the output is shown at regularly
intervals.

Not setting the flush_interval or setting it to a value less than 1
results in the old behavior.